Force response future resolution before inspection
Closes #22.

If the `response` passed to `show()` is not resolved at the time of
invocation (or at least the problem `id` is not assigned yet), then
we're unable to access the captured low-level problem data (i.e.
parameters).

The response is resolved at later point anyway (when samples are
accessed, for example), but it's critical to know the problem id
earlier, when analysing available data in the generic `show()`.